Split corrupt_records
Splits corrupt_record.py (which contained a number of different classes)
into a package where each class is contained within it's own module
