+++ src/crptr/synthetic_populations/corruptor_definitions/death_corruptor.py
@@ -1,4 +1,4 @@
-from crptr import corrupt_record
+from crptr.corrupt_records import base
 from crptr.synthetic_populations.corruptor_definitions.base import Corruptors
 
 
@@ -21,25 +21,25 @@ class DeathCorruptors(Corruptors):
         # =====================================================================
         # Record level
         # =====================================================================
-        self.dayMonthSwapDeath = corrupt_record.CorruptSwapAttributes(
+        self.dayMonthSwapDeath = base.CorruptSwapAttributes(
             attr1='day',
             attr2='month',
             attr_name_list=self.columnLabels
         )
 
-        self.deceasedNameSwap = corrupt_record.CorruptSwapAttributes(
+        self.deceasedNameSwap = base.CorruptSwapAttributes(
             attr1='forename(s) of deceased',
             attr2='surname of deceased',
             attr_name_list=self.columnLabels
         )
 
-        self.fatherNameSwap = corrupt_record.CorruptSwapAttributes(
+        self.fatherNameSwap = base.CorruptSwapAttributes(
             attr1='father\'s forename',
             attr2='father\'s surname',
             attr_name_list=self.columnLabels
         )
 
-        self.motherNameSwap = corrupt_record.CorruptSwapAttributes(
+        self.motherNameSwap = base.CorruptSwapAttributes(
             attr1='mother\'s forename',
             attr2='mother\'s maiden surname',
             attr_name_list=self.columnLabels
